requirements.txt
flask
python-dotenv
PyPDF2
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
torch
requests
werkzeug
